Made 13 rooms including a "Win" room and a cliff that you can climb up, but not down
I developed the game map and made it all connect. Added a "win" room so
you can escape the forest you are lost in. Gave the game a vauge story.
Also added a cliff that you can climb up, but you cannot exit back that
way. I also made it so that the back command does not work either once
you climb up it.